requirements.txt
flask-sqlalchemy
PyMySQL
gunicorn
python-dotenv
pytest